already add asessor and kum

+++ database/migrations/2022_02_02_100217_create_kinerjas_table.php
@@ -16,7 +16,7 @@ class CreateKinerjasTable extends Migration
         Schema::create('kinerja', function (Blueprint $table) {
             $table->id();
             $table->string('buktiDokumen');
-            $table->bigInteger('sks');
+            $table->decimal('sks');
             $table->timestamps();
         });
     }
